Enhance validation and documentation across various components
- Added argument validation for chunk sizes in Buffers and PerformanceHelpers.
- Improved XML documentation for FilterByTimeRange and GetTimeRange methods in EventBatchOperations.
- Refined tests for ProcessInChunks and FilterByTimeRange to ensure correct behavior and boundary conditions.
- Updated assertions in PartitionersTests and SpecializedEventStore tests for clarity and accuracy.

+++ src/LockFree.EventStore/LockFreeEventRingBuffer.cs
@@ -110,7 +110,7 @@ public sealed class LockFreeEventRingBuffer
         for (int i = 0; i < batch.Length; i++)
         {
             var position = startTail - batch.Length + i;
-            _buffer[position % _capacity] = batch[i];
+            _buffer[(int)(position % _capacity)] = batch[i];
         }
         
         // Advance head if necessary to maintain ring buffer invariant
